+++ src/components/Main/Metrics/MetricsContent/MetricsItem/MetricsItem.jsx
@@ -9,16 +9,16 @@ import {
   $MetricsItem,
   $Summary,
   $Title,
+  $ErrorLabel,
 } from "./MetricsItem.styles.jsx";
 import {
   MetricsContext,
   STEPS,
 } from "../../../../../store/metrics-context.jsx";
-
-const TOOLTIPS = {
-  carbon: "Limit, current usage and balance of Carbon in kg",
-  summary: "Carbon usage in %",
-};
+import {
+  apiCallToGetEntityChildsCapacity,
+  apiCallToGetTotalCarbonSum,
+} from "../../../../../api/Api.jsx";
 
 function descendantTooltipHandler(currentStep, role) {
   switch (currentStep) {
@@ -34,24 +34,58 @@ function descendantTooltipHandler(currentStep, role) {
 
 export default function MetricsItem({ metric, index, stepInfoHandler }) {
   const { currentStep, stepHandler, thresholds } = useContext(MetricsContext);
-  const [currentUsage, setCurrentUsage] = useState(0);
+  const [childsCapacityAndCarbon, setChildsCapacityAndCarbon] = useState({
+    childsCapacity: 0,
+    currentFootprint: 0,
+  });
+  const [error, setError] = useState(null);
 
   const { id, name, carbonLimit, corporateKey } = metric;
-  const usage = ((currentUsage / carbonLimit) * 100).toFixed(2) + " %";
+  const usage =
+    ((childsCapacityAndCarbon.currentFootprint / carbonLimit) * 100).toFixed(
+      2
+    ) + " %";
   const threshold = carbonBalance();
 
   useEffect(() => {
-    // TO DO strzal do bazy po current usage dla company/tribe etc
-    const usage = 90;
-    setCurrentUsage(usage);
-  }, []);
+    setError(null);
 
-  function descendantsCount(id) {
-    // w zaleznosci co ma zwrocic czy areas, tribes... inne dla employee
-    // TO DO api call to get numebr of areas where company.id = id
-    // na razei mock, powinno byc jako useEffect
-    return 12;
-  }
+    async function fetchData() {
+      try {
+        const childsCapacity = await apiCallToGetEntityChildsCapacity(
+          STEPS[currentStep].stepName.toLowerCase(),
+          metric.id,
+          metric.name
+        ).catch((error) => {
+          setChildsCapacityAndCarbon({
+            ...childsCapacityAndCarbon,
+            childsCapacity: "err",
+          });
+          throw error;
+        });
+
+        const carbonFootprint = await apiCallToGetTotalCarbonSum(
+          STEPS[currentStep].stepName.toLowerCase(),
+          metric.id
+        ).catch((error) => {
+          setChildsCapacityAndCarbon({
+            ...childsCapacityAndCarbon,
+            currentFootprint: -1,
+          });
+          throw error;
+        });
+
+        setChildsCapacityAndCarbon({
+          childsCapacity: childsCapacity,
+          currentFootprint: carbonFootprint[0],
+        });
+      } catch (error) {
+        setError(error);
+      }
+    }
+
+    fetchData();
+  }, [metric]);
 
   function roleAsDescendant() {
     return metric.role && metric.role.name.charAt(0);
@@ -60,9 +94,7 @@ export default function MetricsItem({ metric, index, stepInfoHandler }) {
   function carbonBalance() {
     const balance = parseFloat(usage);
 
-    const threshold = thresholds.findIndex(
-      (element) => balance > element
-    );
+    const threshold = thresholds.findIndex((element) => balance > element);
 
     return threshold;
   }
@@ -97,7 +129,7 @@ export default function MetricsItem({ metric, index, stepInfoHandler }) {
         >
           <p>
             {currentStep < 4
-              ? descendantsCount(id)
+              ? childsCapacityAndCarbon.childsCapacity
               : roleAsDescendant()}
           </p>
           <Tooltip id={"descendant_tooltip_" + index} />
@@ -105,7 +137,7 @@ export default function MetricsItem({ metric, index, stepInfoHandler }) {
         <$Summary
           $threshold={threshold}
           data-tooltip-id={"summary_tooltip_" + index}
-          data-tooltip-content={TOOLTIPS.summary}
+          data-tooltip-content={"Carbon usage in %"}
           data-tooltip-delay-show={1000}
           data-tooltip-place={"right"}
         >
@@ -114,18 +146,44 @@ export default function MetricsItem({ metric, index, stepInfoHandler }) {
         </$Summary>
       </$Content>
 
-      <$Carbon
-        $threshold={threshold}
-        data-tooltip-id={"carbon_tooltip_" + index}
-        data-tooltip-content={TOOLTIPS.carbon}
-        data-tooltip-delay-show={1000}
-        data-tooltip-place={"bottom"}
-      >
-        <div>{carbonLimit}</div>
-        <div>{currentUsage}</div>
-        <div>{carbonLimit - currentUsage}</div>
-        <Tooltip id={"carbon_tooltip_" + index} />
+      <$Carbon $threshold={threshold}>
+        <a
+          data-tooltip-id={"carbon_tooltip_limit" + index}
+          data-tooltip-content={"Carbon limit [kg]"}
+          data-tooltip-delay-show={1000}
+          data-tooltip-place={"bottom"}
+        >
+          {carbonLimit}
+          <Tooltip id={"carbon_tooltip_limit" + index} />
+        </a>
+        <a
+          data-tooltip-id={"carbon_tooltip_current" + index}
+          data-tooltip-content={
+            "Current carbon footprint [" +
+            childsCapacityAndCarbon.currentFootprint +
+            " kg]"
+          }
+          data-tooltip-delay-show={1000}
+          data-tooltip-place={"bottom"}
+        >
+          {Number(childsCapacityAndCarbon.currentFootprint).toFixed(3)}
+          <Tooltip id={"carbon_tooltip_current" + index} />
+        </a>
+        <a
+          data-tooltip-id={"carbon_tooltip_bilance" + index}
+          data-tooltip-content={
+            "Carbon footprint bilance [" +
+            (carbonLimit - childsCapacityAndCarbon.currentFootprint) +
+            " kg]"
+          }
+          data-tooltip-delay-show={1000}
+          data-tooltip-place={"bottom"}
+        >
+          {(carbonLimit - childsCapacityAndCarbon.currentFootprint).toFixed(3)}
+          <Tooltip id={"carbon_tooltip_bilance" + index} />
+        </a>
       </$Carbon>
+      {error !== null ? <$ErrorLabel>{error.message}</$ErrorLabel> : null}
     </$MetricsItem>
   );
 }